Moved concrete implementations from HttpKit to BrowserNetworking framework
