+++ catowser/BrowserNetworking/BrowserNetworking.h
@@ -1,0 +1,19 @@
+//
+//  BrowserNetworking.h
+//  BrowserNetworking
+//
+//  Created by Ermoshin Andrey on 01.07.2021.
+//  Copyright © 2021 andreiermoshin. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for BrowserNetworking.
+FOUNDATION_EXPORT double BrowserNetworkingVersionNumber;
+
+//! Project version string for BrowserNetworking.
+FOUNDATION_EXPORT const unsigned char BrowserNetworkingVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <BrowserNetworking/PublicHeader.h>
+
+
