+++ catowser/BrowserNetworking/Endpoints/DNSoverHTTPS/GoogleJSONDNSoHTTPsEndpoint+AsyncAwait.swift
@@ -10,7 +10,7 @@
 
 import Foundation
 
-extension HttpKit.Client where Server == HttpKit.GoogleDnsServer {
+extension HttpKit.Client where Server == GoogleDnsServer {
     @available(swift 5.5)
     @available(macOS 12, iOS 15.0, tvOS 15.0, watchOS 8.0, *)
     func aaGetIPaddress(ofDomain domainName: String) async throws -> HttpKit.GoogleDNSOverJSONResponse {
