+++ catowser/BrowserNetworking/Endpoints/Google/GoogleSearchEndpoint+AsyncAwait.swift
@@ -8,9 +8,9 @@
 
 #if swift(>=5.5)
 
-import Foundation
+import HttpKit
 
-extension HttpKit.Client where Server == HttpKit.GoogleServer {
+extension HttpKit.Client where Server == GoogleServer {
     @available(swift 5.5)
     @available(macOS 12, iOS 15.0, tvOS 15.0, watchOS 8.0, *)
     public func aaGoogleSearchSuggestions(for text: String) async throws -> HttpKit.GoogleSearchSuggestionsResponse {
