+++ catowser/BrowserNetworking/SearchEngines/GoogleSearchEngine.swift
@@ -8,15 +8,15 @@
 
 import Foundation
 
-extension HttpKit.SearchEngine {
-    public static func googleSearchEngine() -> HttpKit.SearchEngine {
+extension SearchEngine {
+    public static func googleSearchEngine() -> SearchEngine {
         let items: [URLQueryItem] = [
             URLQueryItem(name: "ie", value: "utf-8"),
             URLQueryItem(name: "oe", value: "utf-8"),
             URLQueryItem(name: "client", value: "firefox")
         ]
         
-        return HttpKit.SearchEngine(shortName: "Google",
+        return SearchEngine(shortName: "Google",
                                     domainName: "www.google.com",
                                     path: "search",
                                     queryItems: items,
