+++ catowser/HttpKit/HttpKitHeaders.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 extension HttpKit {
-    enum HttpHeader {
+    public enum HttpHeader {
         case contentType(ContentType)
         case contentLength(Int)
         case accept(ContentType)
@@ -43,7 +43,7 @@ extension HttpKit {
         }
     }
     
-    enum ContentType: String {
+    public enum ContentType: String {
         case json = "application/json"
         /// The following type is used to indicate that the response will contain search suggestions.
         /// Link: [doc](http://www.opensearch.org/Specifications/OpenSearch/Extensions/Suggestions/1.0)
