+++ catowser/catowser/WebSite/WebViewsReuseManager.swift
@@ -9,7 +9,7 @@
 import UIKit
 import CoreBrowser
 import JSPlugins
-import HttpKit
+import BrowserNetworking
 
 /// The class to control memory usage by managing reusage of web views
 final class WebViewsReuseManager {
